# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Spartan Scraper API
  version: 1.0.0-rc1
servers:
- url: http://localhost:8741
security:
- ApiKeyAuth: []
paths:
...
  "/v1/auth/profiles":
    get:
      summary: List auth profiles
      responses:
        '200':
          description: Auth profiles
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AuthProfilesResponse"
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
...
  "/v1/auth/import":
    post:
      summary: Import auth vault
      description: |
        Import auth vault from a file in the data directory.
        The path must be a simple filename (no directory separators or path traversal).
        The file is resolved relative to the configured DATA_DIR.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AuthVaultImport"
      responses:
        '200':
          description: Imported
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StatusResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/v1/auth/export":
    post:
      summary: Export auth vault
      description: |
        Export auth vault to a file in the data directory.
        The path must be a simple filename (no directory separators or path traversal).
        The file is resolved relative to the configured DATA_DIR.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AuthVaultExport"
      responses:
        '200':
          description: Exported
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StatusResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/v1/auth/oauth/initiate":
    post:
      summary: Initiate OAuth 2.0 authorization flow
      description: |
        Starts an OAuth 2.0 authorization code flow. Generates state parameter for CSRF protection
        and optionally PKCE code challenge. Returns the authorization URL to redirect the user to.
      operationId: initiateOAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/OAuthInitiateRequest"
      responses:
        '200':
          description: Authorization URL generated
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/OAuthInitiateResponse"
        '400':
          description: Bad Request - Missing profile or invalid OAuth configuration
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/v1/auth/oauth/callback":
    get:
      summary: OAuth 2.0 callback endpoint
      description: |
        Handles the OAuth 2.0 callback from the authorization server. Exchanges the authorization
        code for access and refresh tokens. The state parameter is validated for CSRF protection.
      operationId: oauthCallback
      parameters:
      - name: code
        in: query
        description: Authorization code from OAuth provider
        schema:
          type: string
      - name: state
        in: query
        required: true
        description: State parameter for CSRF validation
        schema:
          type: string
      - name: error
        in: query
        description: Error from OAuth provider (if authentication failed)
        schema:
          type: string
      responses:
        '200':
          description: Token exchange successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/OAuthTokenResponse"
        '400':
          description: Bad Request - Invalid state or missing code
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/v1/auth/oauth/refresh":
    post:
      summary: Refresh OAuth 2.0 access token
      description: |
        Refreshes an OAuth 2.0 access token using the stored refresh token.
        The new tokens are automatically saved for the profile.
      operationId: refreshOAuthToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/OAuthRefreshRequest"
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/OAuthTokenResponse"
        '400':
          description: Bad Request - Missing profile or no refresh token
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Profile or token not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/v1/auth/oauth/discover":
    post:
      summary: Perform OIDC discovery
      description: |
        Fetches OIDC provider metadata from a discovery URL or issuer.
        Returns endpoints, supported scopes, and other provider configuration.
      operationId: discoverOIDC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/OIDCDiscoverRequest"
      responses:
        '200':
          description: Discovery successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/OIDCProviderMetadata"
        '400':
          description: Bad Request - Missing discovery URL or issuer
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/v1/auth/oauth/revoke":
    post:
      summary: Revoke OAuth 2.0 token
      description: |
        Revokes the OAuth 2.0 token for a profile at the provider (if supported)
        and deletes it from local storage.
      operationId: revokeOAuthToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - profile_name
              properties:
                profile_name:
                  type: string
                  description: Name of the profile to revoke tokens for
      responses:
        '200':
          description: Token revoked successfully
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StatusResponse"
        '400':
          description: Bad Request - Missing profile name
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Profile or token not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/v1/templates":
    get:
      summary: List extraction templates
      description: Returns list of all available extraction templates (built-in +
        file-based)
      operationId: listTemplates
      responses:
        '200':
          description: Template list
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items:
                      type: string
                    description: List of template names
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
    post:
      summary: Create a new template
      description: Create a custom extraction template
      operationId: createTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateTemplateRequest"
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TemplateDetail"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
...
  "/v1/template-preview":
    get:
      summary: Fetch page DOM structure for selector building
      description: Fetches a remote page and returns its simplified DOM tree for visual
        selector building
      operationId: getTemplatePreview
      parameters:
      - name: url
        in: query
        required: true
        schema:
          type: string
          format: uri
        description: URL to fetch and analyze
      - name: headless
        in: query
        schema:
          type: boolean
          default: false
        description: Use headless browser for fetching
      - name: playwright
        in: query
        schema:
          type: boolean
          default: false
        description: Use Playwright instead of Chromedp for headless fetching
      responses:
        '200':
          description: DOM preview
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TemplatePreviewResponse"
        '400':
          description: Bad Request - invalid URL
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/v1/template-preview/test-selector":
    post:
      summary: Test CSS selector against URL
      description: Tests a CSS selector against a fetched page and returns matching
        elements
      operationId: testSelector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TestSelectorRequest"
      responses:
        '200':
          description: Selector test results
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TestSelectorResponse"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/v1/ai/extract-preview":
    post:
      summary: Preview AI-powered extraction
      description: Extract structured data using AI/LLM without creating a job.
        Provide HTML directly or let Spartan fetch the URL through its existing
        fetch stack.
      operationId: aiExtractPreview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AIExtractPreviewRequest"
      responses:
        '200':
          description: AI extraction preview results
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AIExtractPreviewResponse"
        '400':
          description: Bad Request - invalid parameters
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
  "/v1/ai/template-generate":
    post:
      summary: Generate extraction template using AI
      description: Generate an extraction template using AI from either fetched page content
        or raw HTML supplied directly.
      operationId: aiTemplateGenerate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AIExtractTemplateGenerateRequest"
      responses:
        '200':
          description: Generated template
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AIExtractTemplateGenerateResponse"
        '400':
          description: Bad Request - invalid parameters
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Generate via CLI with `spartan auth
        apikey generate`.
    SessionAuth:
      type: apiKey
      in: cookie
      name: spartan_session
      description: Session cookie for web UI authentication
  schemas:
...
    OAuthInitiateRequest:
      type: object
      required:
      - profile_name
      properties:
        profile_name:
          type: string
          description: Name of the profile with OAuth2 configuration
        redirect_uri:
          type: string
          format: uri
          description: Override redirect URI for this flow
    OAuthInitiateResponse:
      type: object
      properties:
        authorization_url:
          type: string
          format: uri
          description: URL to redirect user for OAuth authorization
        state:
          type: string
          description: CSRF protection state parameter
...
    OAuthRefreshRequest:
      type: object
      required:
      - profile_name
      properties:
        profile_name:
          type: string
          description: Name of the profile to refresh token for
...
    AuthProfilesResponse:
      type: object
      properties:
        profiles:
          type: array
          items:
            "$ref": "#/components/schemas/AuthProfile"
    AuthVaultImport:
      type: object
      properties:
        path:
          type: string
          pattern: "^[^/\\\\]+$"
          description: Simple filename only (no path, no directory separators, resolved
            relative to DATA_DIR)
      required:
      - path
    AuthVaultExport:
      type: object
      properties:
        path:
          type: string
          pattern: "^[^/\\\\]+$"
          description: Simple filename only (no path, no directory separators, resolved
            relative to DATA_DIR)
      required:
      - path
    StatusResponse:
      type: object
      properties:
        status:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
      required:
      - error
...